+++ R/clusterSimmatrix.R
@@ -47,21 +47,22 @@ clusterSimmatrix <- function(simmatrix,
                              dist_mthd = "correlation", hclust_mthd = "average",
                              nBoot = 1000,  nClusters = FALSE, r=seq(.7,1.4,by=.1)){
 
-  if("pvclust" %in% rownames(utils::installed.packages()) == FALSE)
-    stop("This function applies library 'pvclust' for parallel processing.
-         Please install this package.")
+  if (!requireNamespace("pvclust")) {
+    stop(" The \"pvclust\" package is required for parallel processing.")
+  }
 
   if(class(simmatrix) != "matrix") stop("Object 'x' is not of class 'Matrix'")
   if(nClusters != FALSE){
 
     if(!is.numeric(nClusters))
       stop("nClusters must be set to 'FALSE' or a numeric value.")
-    if("parallel" %in% rownames(utils::installed.packages()) == FALSE)
-      stop("This function applies library 'parallel' for parallel processing.
-           Please install this package.")
-    if("doParallel" %in% rownames(utils::installed.packages()) == FALSE)
-      stop("This function applies library 'doParallel' for parallel processing.
-           Please install this package.")
+
+    if (!requireNamespace("parallel")) {
+      stop(" The \"parallel\" package is required for parallel processing.")
+    }
+    if (!requireNamespace("doParallel")) {
+      stop(" The \"doParallel\" package is required for parallel processing.")
+    }
 
     cl <- parallel::makeCluster(nClusters)
     doParallel::registerDoParallel(cl)